requirements.txt
streamlit
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
pillow
scikit-learn==1.0.2
keras
tensorflow